requirements.txt
# NS-MAS Requirements
# Note: Clingo must be installed via conda: conda install -c conda-forge clingo=5.8.0

# Core ML/AI
langgraph>=1.0.0
langchain>=1.0.0
langchain-openai>=0.1.0
langchain-anthropic>=0.1.0
langsmith>=0.1.0

# Contextual Bandit
vowpalwabbit>=9.10.0

# ASP Interface
clorm>=1.6.0

# Data Science
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# NLP/Embeddings
sentence-transformers>=2.2.0

# Utilities
python-dotenv>=1.0.0
tqdm>=4.65.0
aiohttp>=3.8.0

# Testing
pytest>=7.0.0
